Remove HTML template version of the GradeMe Button
We've been using the XBlock version of this component for a few years
now. The XBlock allows us to make changes to all existing instances of
the problem, whereas the HTML template requires manual modifications to
be made to each and every component.

We will schedule follow-up work to:
- query the modulestore for existing templated instances
  (via coursegraph/neo4j)
- Convert existing instances to the XBlock version

The follow-up work could be completed at any time (or technically, never).
Since this is an HTML template, the template itself can be removed
without effecting existing instances.
This commit prevents any new instances from being created.