Fixed drop down buttons on header of bulk settings page
I don't know why this line importing jquery was included here because jquery is usually automatically loaded. However, when I removed this line, it fixes the bug with the drop down buttons on the bulk settings page.
I think it may have to do with the order of the javascript loading, and was likely introduced after pulling from Gingko.